[hir2mpl]Internal commit msg:     __builtin_prefetch bugfix     [no-builtin] add support for -fno-builtin option     [cg] support prefetch, clear_cahce, unreachable builtin     [hir2mpl] support builtin_constant_p     [cg] support __builtin_alloca_with_align     [hir2mpl] support pragma preferinline     fe_bugfix_ExprStmt_scope_fix     Support hot attr and prop temperature across callgraph (not enable yet), refactor funcAttr     [mapleall] vrp fix shortcircut opt     [driver] file type     [driver]deal maple test.c ../test.c     [armneon] improve inline prop for arm neon functions     [hir2mpl]fix DesignatedInitUpdaterExpr initialization in constantArray [mapleall]Internal commit msg:     [mapleall] bugfix vrp     [cg] cleancode     [driver]code style     [mplcg] fix memory overlap bug for CombineContiLoadAndStorePattern     __builtin_prefetch bugfix     fix implicit cvt in simplify     [no-builtin] add support for -fno-builtin option     [cg] pgo use bugfix     [mplme] outline bugfix     cg_bugfix_cleancode_4243900731     [driver]code style     clean code     cleancode     [mapleall] cleancode2  64     cleancode     global opt bug     [cg] support prefetch, clear_cahce, unreachable builtin     [cg] cleancode     [driver]code style     [cg] clean code raopt/sink     [driver]code style     [driver] filtering maple option when call gcc     [mplcg] clean code     [cg] support __builtin_constant_p     [cg] bugfix for implicit cvt in regassign     [mplme] clean code     [mapleall] bugfix objsize compute last field size of agg     link option type judgment     [driver] spec ref os     [mapleall] bugfix deal with cvt in vrp     [cg] tbz pattern del edge bugfix     [mapeall] Inlining tuning according to frequency and unlikely     [cg]Enhance ICOSameCondPattern     cg_bugfix_fix_add_cmp_zero_pattern_check_condition     [mplcg] add RSPattern in validbitopt phase     [cg] OP_select implicit cvt bugfix     [cg] delCbzCbnz need check target is nextBB     [mplcg] support deep copy in CG and fix DuplicateBBPattern bug     [driver]fixbug option whith whiteSpace     [cg] support __builtin_alloca_with_align     cg_bugfix_fix_coldhot_split_startend_label     do not sink stmt which contains volatile exprs     cleancode     [cg] fix chaining pattern runtime bug in cfgo     [mapleall] bugfix deal with iread     [driver] specs option     [driver] check option vector is empty     clean code     [mplme] add no ipaclone func list option     [cg]clean code     [cg] bugfix implicit conversion in handlefunc     [driver] change fPIE + fpie combination strategy     uint32 reverse bugfix     fix constant overflow check bug     [driver]fixbug DTS2023072502356     [driver] duplicate file separator     [ICSL] Clean Code 69     cleancode     [cg]Enhance ico for issue 239     target prop bug     fix ebo bug     [hir2mpl] support pragma preferinline     cg_bugfix_debug_testcase     [ci] split pgo cases     [driver]Lto Option     jump pad     [mplcg] change Lo12li verify condition     [driver] fake.a     Support hot attr and prop temperature across callgraph (not enable yet), refactor funcAttr     [mplcg] clean code about scheduling     get opnd type from `opndtype` instead of `primtype` of OP_cvt     clean code     prefer_inline pragma implemention     [cleancode] maple driver     [cg] optimize add insn in selecting rangegoto     [cg] tbz pattern in peep bugfix - targetBB is ftBB     [cg] enable post ra sink     [driver] link gcc option     [mapleall] enhance opt continuous memory     me_bugfix_fix_add_scope_boundary_condition     [cg] eliminate redundant sbfx instruction in validbit opt phase     [mplcg] do not combine x16 in AddSubMergeLdStPattern of cgpostpeephole     [cg] bugfix for preg copy when has differ SSA version     [mplme] bugfix of cfgo     [mpl2mpl] change MustInline condition for external gnu inline func     clean code for copy ret     cleancode     [mapleall] vrp fix shortcircut opt     [litepgo] hpf2 daily solution for cold entry bb     [driver] file type     cg_bugfix_tls_testcase_and_fix     [cg]Bugfix for ico     [cg] RA loop split lr has ref in exitBB bugfix     cg_bugfix_tls_not_emit_location     [driver]deal maple test.c ../test.c     cleancode switch lower     cleancode vrp     [ICSL] Clean Code 68     [cg] stack protected sink clean code and add test     [driver]don't pass --save-temps to clang     [mplcg] tuning scheduling v2.0 back to version11     [maplecg]switch lower enhance     me_bugfix_not_check_subscope     cg_bugfix_fix_internalFlag1_count     [mplcg] Modify mov 32-bits immediate valid range     [cg] clean code 0717     [ico]Enhance ICOSameCondPattern for cbz/cbnz     [cg] RA no need remat at spillpost     [armneon] improve inline prop for arm neon functions     [cg] stack protect sink bugfix     [ICSL] Clean Code 67     [mapleall]resolve issue     [mplall] algorithm refactor     cg_bugfix_tls_warmup_fix     [cg] cfg relationship maintenance error when optimizing     code clean     [cg] get field align

+++ src/mapleall/maple_me/src/me_bb_analyze.cpp
@@ -16,7 +16,7 @@
 
 namespace maple {
   void BBAnalyze::SetHotAndColdBBCountThreshold() {
-    std::vector<uint32> times;
+    std::vector<uint64> times;
     auto eIt = cfg->valid_end();
     for (auto bIt = cfg->valid_begin(); bIt != eIt; ++bIt) {
       auto *bb = *bIt;
@@ -27,8 +27,8 @@ namespace maple {
         static_cast<uint32>(static_cast<double>(times.size()) / kPrecision * (MeOption::profileBBHotRate));
     uint32 coldIndex =
         static_cast<uint32>(static_cast<double>(times.size()) / kPrecision * (MeOption::profileBBColdRate));
-    hotBBCountThreshold = times.at(hotIndex);
-    coldBBCountThreshold = times.at(coldIndex);
+    hotBBCountThreshold = static_cast<uint32>(times.at(hotIndex));
+    coldBBCountThreshold = static_cast<uint32>(times.at(coldIndex));
   }
 
   bool BBAnalyze::CheckBBHot(const BBId bbId) const {
